Backend/task_tracker_django/task_tracker/my_serializers.py
from task_tracker import models
from rest_framework import serializers


class TaskTrackerSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.TaskTracker
        fields = '__all__'
